Move contexts out of components folder
