+++ src/pages/our-mission-vision-and-values-page/ourMissionVisionAndValuesPage.test.tsx
@@ -14,7 +14,7 @@ import navigationBarFactory from '../../test/factories/strapi/NavigationBarFacto
 import FooterFactory from '../../test/factories/strapi/FooterFactory';
 import OurMissionVisionAndValuesPageFactory from '../../test/factories/strapi/OurMissionVisionAndValuesPageFactory';
 import OurMissionVisionAndValuesPage from './OurMissionVisionAndValuesPage';
-import { SharedDataContext } from '../../components/contexts/SharedDataProvider';
+import { SharedDataContext } from '../../contexts/SharedDataProvider';
 
 vi.mock('react-router-dom', async () => {
   const actual = await vi.importActual('react-router-dom');
