+++ src/pages/our-team-page/ourTeamPage.test.tsx
@@ -14,7 +14,7 @@ import navigationBarFactory from '../../test/factories/strapi/NavigationBarFacto
 import FooterFactory from '../../test/factories/strapi/FooterFactory';
 import OurTeamPageFactory from '../../test/factories/strapi/OurTeamPageFactory';
 import OurTeamPage from './OurTeamPage';
-import { SharedDataContext } from '../../components/contexts/SharedDataProvider';
+import { SharedDataContext } from '../../contexts/SharedDataProvider';
 
 vi.mock('react-router-dom', async () => {
   const actual = await vi.importActual('react-router-dom');
